move everything into pkg directory
